Move subcommands files
